Use bundled catch.hpp for testing
bdep CI builds were failing in a "test-installed" phase, which checks
whether the project's provided tests can work against an installed
version of the library. This seems like a good thing, because it means
it checks that the install command does the right thing and that the
library is in a useable state for downstream.

The problem seems to be that because catch2 is not a dependency of
the library itself, just the tests, it wasn't (and shouldn't be)
installed -- but this means that the tests can't find the library/header
either, hence the failure.

The recommended way around this (from Boris Kolpakov on Slack) is to
make a separate leftpad-tests package within the parent project, and
make the -tests package depend on catch2. This makes sense, but having
tried to do it, it seems to require more Build2 knowledge than I have
right now to get it right. This is a shame, because I'd imagine that
wanting to use Catch (or GTest or DocTest or whatever) for testing
(only) is a very common thing to want to do, and it *nearly* works in
Build2 but not quite.

Anyway, perhaps I'll have another go at splitting the project up into
separate lib and test packages at some point. Buf for now I'll just
take the easy route and bundle catch2.hpp manually instead :(